Separating logic from UI components
