+++ packages/nextjs-app-ts/src/components/balancer-eth-components/balancer-array/BalancerArray.tsx
@@ -1,38 +1,15 @@
 import { ParamType } from '@ethersproject/abi';
 import { Button } from 'antd';
-import * as _ from 'lodash';
-import { FC, useContext, useState } from 'react';
+import { FC } from 'react';
 
 import { BalancerInput } from '../balancer-input/BalancerInput';
 
+import { useBalancerArray } from './BalancerArray.hook';
 import { IBalancerArray } from './IBalancerArray';
 
-import { BalancerFunctionContext } from '~~/components/balancer-eth-components/balancer-function/BalancerFunction.context';
-
-export const BalancerArray: FC<IBalancerArray> = ({ input, inputIndex }) => {
-  const { setInputValues } = useContext(BalancerFunctionContext);
-
-  const [numberArrayElements, setNumberArrayElements] = useState(0);
-
-  const title = input.name ? `${input.name} (${input.type})` : input.type;
-  const elementsArray = Array.from(Array(numberArrayElements).keys());
-
-  const onAddClick = () => {
-    setNumberArrayElements((old) => old + 1);
-  };
-
-  const onRemoveClick = () => {
-    setInputValues((oldInputValues) => {
-      const newInputValues = [...oldInputValues];
-      const valuesArray = _.get(newInputValues, inputIndex) as unknown as any[];
-      if (valuesArray && valuesArray.length >= numberArrayElements) {
-        valuesArray.pop();
-      }
-      _.set(newInputValues, inputIndex, valuesArray);
-      return newInputValues;
-    });
-    setNumberArrayElements((old) => (old <= 0 ? 0 : old - 1));
-  };
+export const BalancerArray: FC<IBalancerArray> = (props) => {
+  const { input, inputIndex } = props;
+  const { elementsArray, onRemoveClick, onAddClick, title } = useBalancerArray(props);
 
   return (
     <div className={'balancer-array-container'}>
