+++ packages/nextjs-app-ts/src/components/balancer-eth-components/balancer-boolean/BalancerBoolean.tsx
@@ -1,28 +1,12 @@
 import { Checkbox } from 'antd';
-import { CheckboxChangeEvent } from 'antd/es/checkbox';
-import * as _ from 'lodash';
-import { FC, useContext } from 'react';
+import { FC } from 'react';
 
-import { BalancerFunctionContext } from '../balancer-function/BalancerFunction.context';
-
+import { useBalancerBoolean } from './BalancerBoolean.hook';
 import { IBalancerBoolean } from './IBalancerBoolean';
 
-export const BalancerBoolean: FC<IBalancerBoolean> = ({ input, inputIndex }) => {
-  const { inputValues, setInputValues } = useContext(BalancerFunctionContext);
-
-  const title = input.name ? `${input.name} (${input.type})` : input.type;
-  const currentValue = _.get(inputValues, inputIndex) as unknown as boolean;
-
-  const onInputChange = (event: CheckboxChangeEvent) => {
-    const value = event.target.checked;
-    setInputValues((oldInputValues) => {
-      const newInputValues = [...oldInputValues];
-      _.set(newInputValues, inputIndex, value);
-      return newInputValues;
-    });
-  };
-
-  console.log('Here!');
+export const BalancerBoolean: FC<IBalancerBoolean> = (props) => {
+  const { input } = props;
+  const { currentValue, onInputChange, title } = useBalancerBoolean(props);
 
   return (
     <div className={'balancer-input-container'}>
