+++ packages/nextjs-app-ts/src/components/balancer-eth-components/balancer-text/BalancerText.tsx
@@ -1,25 +1,12 @@
 import { Input } from 'antd';
-import * as _ from 'lodash';
-import { ChangeEvent, FC, useContext } from 'react';
-
-import { BalancerFunctionContext } from '../balancer-function/BalancerFunction.context';
+import { FC } from 'react';
 
 import { IBalancerText } from './IBalancerText';
 
-export const BalancerText: FC<IBalancerText> = ({ input, inputIndex }) => {
-  const { inputValues, setInputValues } = useContext(BalancerFunctionContext);
+import { useBalancerText } from '~~/components/balancer-eth-components/balancer-text/BalancerText.hook';
 
-  const title = input.name ? `${input.name} (${input.type})` : input.type;
-  const currentValue = _.get(inputValues, inputIndex) as unknown as string;
-
-  const onInputChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value;
-    setInputValues((oldInputValues) => {
-      const newInputValues = [...oldInputValues];
-      _.set(newInputValues, inputIndex, value);
-      return newInputValues;
-    });
-  };
+export const BalancerText: FC<IBalancerText> = (props) => {
+  const { currentValue, onInputChange, title } = useBalancerText(props);
 
   return (
     <div className={'balancer-input-container'}>
